requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
et_xmlfile==2.0.0
filelock==3.18.0
idna==3.10
numpy==2.2.6
openpyxl==3.1.5
pandas==2.2.3
pyarrow==21.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
six==1.17.0
tzdata==2025.2
urllib3==2.4.0
vulture==2.14
